mm/damon/core: implement damon_initialized() function
Patch series "mm/damon: define and use DAMON initialization check
function".

If DAMON is used when it is not yet successfully initialized, the caller
could crash.  The DAMON core layer does not providing a reliable way to
see if it is successfully initialized and therefore ready to be used.  As
a result, DAMON API callers are implementing their own hacks to see this. 
The hacks simply assume DAMON should be ready on module init time.  It is
not reliable as DAMON initialization can indeed fail if KMEM_CACHE()
fails, and difficult to maintain as those are duplicates.  Implement a
core layer API function for better reliability and maintainability to
replace the hacks with followup commits.


This patch (of 7):

If DAMON is tried to be used when it is not yet successfully initialized,
the caller could be crashed.  DAMON core layer is not providing a reliable
way to see if it is successfully initialized and therefore ready to be
used, though.  As a result, DAMON API callers are implementing their own
hacks to see it.  The hacks simply assume DAMON should be ready on module
init time.  It is not reliable as DAMON initialization can indeed fail if
KMEM_CACHE() fails, and difficult to maintain as those are duplicates. 
Implement a core layer API function for better reliability and
maintainability to replace the hacks with followup commits.

Link: https://lkml.kernel.org/r/20250916033511.116366-2-sj@kernel.org
Link: https://lkml.kernel.org/r/20250916033511.116366-2-sj@kernel.org
Signed-off-by: SeongJae Park <sj@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>